feature use ng-router-loader refactor(+detail): use as barrel
