+++ package.json
@@ -98,7 +98,6 @@
     "@types/source-map": "^0.5.0",
     "@types/uglify-js": "^2.0.27",
     "@types/webpack": "^2.0.0",
-    "angular-router-loader": "^0.4.0",
     "angular2-template-loader": "^0.6.0",
     "assets-webpack-plugin": "^3.4.0",
     "awesome-typescript-loader": "~3.0.0-beta.17",
@@ -123,6 +122,7 @@
     "karma-remap-coverage": "^0.1.4",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "1.8.1",
+    "ng-router-loader": "^1.0.0",
     "ngc-webpack": "^1.0.2",
     "node-sass": "^4.1.1",
     "npm-run-all": "^4.0.0",
