+++ src/app/+detail/detail.routes.ts
@@ -1,0 +1,5 @@
+import { DetailComponent } from './detail.component';
+
+export const routes = [
+  { path: '', component: DetailComponent, pathMatch: 'full' },
+];
