+++ src/app/+detail/index.ts
@@ -1,26 +1,1 @@
-import { CommonModule } from '@angular/common';
-import { DetailComponent } from './detail.component';
-import { FormsModule } from '@angular/forms';
-import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-
-console.log('`Detail` bundle loaded asynchronously');
-// async components must be named routes for WebpackAsyncRoute
-export const routes = [
-  { path: '', component: DetailComponent, pathMatch: 'full' },
-];
-
-@NgModule({
-  declarations: [
-    // Components / Directives/ Pipes
-    DetailComponent,
-  ],
-  imports: [
-    CommonModule,
-    FormsModule,
-    RouterModule.forChild(routes),
-  ],
-})
-export class DetailModule {
-  public static routes = routes;
-}
+export { DetailModule } from './detail.module';
