+++ src/app/app.routes.ts
@@ -9,6 +9,6 @@ export const ROUTES: Routes = [
   { path: '',      component: HomeComponent },
   { path: 'home',  component: HomeComponent },
   { path: 'about', component: AboutComponent },
-  { path: 'detail', loadChildren: './+detail/index#DetailModule'},
+  { path: 'detail', loadChildren: './+detail#DetailModule'},
   { path: '**',    component: NoContentComponent },
 ];
